Fix channel quantity "non-zero" bug
The initialisation of the minChannelSizes variable did not
differentiate between the original and modelled
(channelname_prime) channels and would thus calculate min
sizes for every channel. While these values were only used
to set the channel sizes of the modelled channels, it did
lead to a distribution size that was double of what it had
to be. This was fixed by setting all original channel sizes
in the minChannelSizes map to 0.

This had the side-effect of uncovering another bug, where
graphs with initial tokens were being left with a non-zero
value. Setting the preloads of dataflow_prime therefore now
sets all channel preloads according to minChannelSizes
regardless of whether or not they are modelled channels or
not.