Track token prod/cons rate using execution count
Previously tracked token production/consumption rate by keeping
track of the phase the actor was in relative to each port. To
simplify this, we instead use the number of executions to
determine the correct token production/consumption rate by
taking the mod of the number of executions and the phase count
of each port.