Separate start and end execution functions
In order to account for variable execution times, the start and
end of executions need to be defined in separate functions. At
the start of an actor execution, it will consume tokens; the
production of tokens should only occur after it's corresponding
execution time - the tokens that it would produce should not be
available before it has completed executing.