Add alternate getter for current actor phase
Added a function to get the phase of the actor that doesn't rely
on the phase count of a particular port. The use of this assumes
that actors have the same number of phases of execution for all
ports. Previously relied on a function that relied on the phase
count of a port (associated with an edge) in order to allow for
actors with a variable number of phases for different ports.