Track execution time and phase in execution queue
Implement the execution queue in the state class as a list of pairs
containing the remaining execution time of each actor firing as well
as the phase of execution that item in the list corresponds to. Also
implemented functions to remove latest executions from queue as well
as checking when an actor is ready to end executions.