Add Actor ID functions
Add field 'id' to actor implementation and related methods.

We need IDs to easily compare actors identified for certain attributes.
This will prove useful to identify the actor with the lowest repetition
factor in order to track their executions. This is how we will determine
when to store a state during ASAP execution.