Add throughput computation implementation
Added throughput computation function in StateList class.

No real reason why it was implemented in state.cpp rather than
in symbolic_execution.cpp. Might want to think about most appropriate
location in future updates.

The state list stores the iterator of added states that are already
stored in it (representing a revisited state). By iterating from the
first occurance of the revisited state to the end of the list, we
can compute the number of iterations and the total time elapsed -
the ratio of the two gives us the throughput.