Add helper function for SCC detection
Added new class (scc.cpp) to handle SCC detection. Helper function
implemented to compute the order in which the DFS should be computed
in the transpose of the graph.

This will allow us to determine the SCC of the graph according to
Kosaraju's algorithm.

computeDFSStack will perform a DFS of the input graph, starting at
the actor corresponding to 'startId'. After the DFS completes visiting
all of each actor's neighbour cells, the actor's ID will be pushed
onto the 'dfsOrder' stack.