Add function to identify separate SCC
Add functions to identify SCC based on order stack from
computeDFSStack.

By having a function that returns a map containing separate SCCs,
we can then separate the choice of SCC identification algorithm from
the function of actually generating the SCC as a Dataflow subgraph.

In the order given by the stack of actor IDs from computeDFSStack,
computeSCCKosaraju performs a DFS on the transpose of the original
graph by calling dfsTranspose. The actors visited by dfsTranspose are
pushed to a vector, which is then stored in a unique key in a map sccMap
according to the component that they belong to.