Fix bug in DFS implementation
Iterated through the list of 'visitedActors' and applied recursive DFS
to fill order stack. Prior to this, the DFS was just called on an
arbitrary "first" actor, which led to cases where some actors would
not be visited after the DFS --- e.g. bipartite.xml. This would then
lead to an incorrect SCC identification in subsequent steps.