Fix reentrancy workaround
Accounts for the number of phases when computing the throughput of a
standalone actor.

Prior to this, we were simply assuming that the
throughput of the standalone component (consisting of a single actor
with a presumably ignored re-entrant edge) was 1. The scaled
throughput was then computed by dividing that by the actor's
repetition factor from the original dataflow graph. This returned the
wrong throughput when this actor had more than one phase.

We now multiply the standalone throughput by it's phase count before
dividing it by it's repetition factor from within the original
dataflow.